v2  step 07  update

+++ pages/workouts/[id].jsx
@@ -1,18 +1,21 @@
+import Link from "next/link";
 import { useRouter } from "next/router";
 import styled from "styled-components";
 import useSWR from "swr";
 import { collectMuscleGroups } from "@/lib/utils";
+import { useState } from "react";
 export default function WorkoutDetailsPage() {
+  const [isDeleting, setIsDeleting] = useState(false);
   const router = useRouter();
   const id = router.query.id;
   const apiUrl = `/api/workouts/${id}`;
 
   const { data: workout, isLoading, error } = useSWR(apiUrl);
   if (isLoading) {
-    <main>loading</main>;
+    return <main>loading</main>;
   }
   if (error) {
-    <main>error</main>;
+    return <main>error</main>;
   }
   if (!workout) {
     return <main>loading workout</main>;
@@ -21,11 +24,15 @@ export default function WorkoutDetailsPage() {
   const muscleGroupsInWorkout = collectMuscleGroups(exercises);
 
   const handleDelete = async () => {
+    if (!isDeleting) {
+      setIsDeleting(true);
+      return;
+    }
     const response = await fetch(apiUrl, {
       method: "DELETE",
     });
     if (response.ok) {
-      router.push(`/`);
+      router.push(`/workouts`);
     }
   };
 
@@ -37,9 +44,9 @@ export default function WorkoutDetailsPage() {
       <ul>
         {exercises.map(({ reps, sets, exercise }) => {
           return (
-            <li key={exercises.name}>
+            <li key={exercise.name}>
               <StyledSpan>{exercise.name} </StyledSpan>
-              || Reps: {reps} - Sets: {sets}
+              _Reps: {reps} - Sets: {sets}
             </li>
           );
         })}
@@ -50,14 +57,39 @@ export default function WorkoutDetailsPage() {
           return <li key={muscle}>{muscle}</li>;
         })}
       </ul>
-      <button type="button" onClick={handleDelete}>
-        Delete
+      <button
+        type="button"
+        onClick={() => {
+          router.push(`/workouts/update/${id}`);
+        }}
+      >
+        update
       </button>
+      {isDeleting ? (
+        <div>
+          <button type="button" onClick={() => setIsDeleting(false)}>
+            Cancel Delete
+          </button>
+          <button type="button" onClick={handleDelete}>
+            Confirm Deletion
+          </button>
+        </div>
+      ) : (
+        <div>
+          <button type="button" onClick={handleDelete}>
+            Delete
+          </button>
+        </div>
+      )}
+      <div>
+        <Link href={`/`}>home</Link>
+        <Link href={`/workouts`}>Workouts</Link>
+      </div>
     </main>
   );
 }
 
 const StyledSpan = styled.span`
   display: inline-block;
-  width: 120px;
+  width: 160px;
 `;
